LC-find-xor-sum-of-all-pairs-bitwise-and [Runtime: 41 ms; Memory: 97.3 MB]

+++ bit-manipulation/find-xor-sum-of-all-pairs-bitwise-and.cpp
@@ -22,7 +22,7 @@ public:
                 }
             }
 
-            long long x=n_set*m_set;
+            long long x=(long long)n_set*m_set;
             if(x%2==1){
                 xor_v[i]=1;
             }
